+++ seedfarmer/commands/__init__.py
@@ -14,6 +14,7 @@
 
 from seedfarmer.commands._deployment_commands import apply, destroy
 from seedfarmer.commands._module_commands import deploy_module, destroy_module
+from seedfarmer.commands._parameter_commands import generate_export_env_params
 from seedfarmer.commands._stack_commands import (
     deploy_managed_policy_stack,
     deploy_module_stack,
@@ -30,4 +31,5 @@ __all__ = [
     "deploy_module_stack",
     "destroy_module_stack",
     "deploy_seedkit",
+    "generate_export_env_params",
 ]